add ability to enter functions in REPL
this works by checking whether there are unmatched brackets in the
input. if there are, the input is appended to a buffer and the next line
is read. this is repeated until all brackets are matched.

so only the first one would work if entered line by line into the repl

```
fn add(a, b) {     <-- unmatched left brace
    a + b
}
```

```
fn add(a, b)       <-- all brackets are matched, so this is evaluated as
{                      is, resulting in an error
    a + b
}
```