fix crucial bug in parser related to operator precedence
The problem was that was that the previous implementation would only
build up a chain of operators that increase in precedence. Once meeting
an operator of lower precedence than the previous one, it would go up
the whole chain and use the current chain as the left hand side of that
operator. This would lead to the following AST when parsing the
expression `1 + 2 ^ 3 * 4`:

            *       <-- The problem is that the `*` operator has lower
      +       4     precedence than the `^` operator which then puts the
    1   ^           `*` operator on top of the current chain. This is
      2   3         obviously wrong, since it still has higher
                    precedence than the `+` operator.

The fixed implementation is able to "walk up the chain" when meeting an
operator of lower precedence, instead of jumping to the top right away.
Thus it is able to insert the operator in the right place, which results
in the following AST for the expression above:

      +
    1         *     <-- The `*` operator is now correctly inserted here,
          ^     4   since it has higher precedence than the `+`
        2   3       operator. This is detected while walking up the
                    chain.